Fixes bug where transactions cannot be decoded
The bitcoind RPC interface doesn't send a "hex" field at least since
version 0.7. This patch makes the parsing of the "hex" field optional,
and in either case, always stores the raw transaction object; this is to
remain API compatible. In long term, we might want to get rid of the
decRaw field of DecodedRawTransaction since it is redundant.